requirements.txt
Flask
Flask-SQLAlchemy
Authlib
requests
dnspython
requests-doh
python-dotenv
web3
eth-account